Solaris gcc -m64 fixes
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@11930 06827809-a52a-0410-b366-d66718629ded

+++ agent/helpers/watcher.c
@@ -217,7 +217,7 @@ netsnmp_watched_timestamp_handler(netsnmp_mib_handler *handler,
                                netsnmp_request_info *requests)
 {
     marker_t timestamp = (marker_t) handler->myvoid;
-    int      uptime;
+    long     uptime;
     int      cmp;
 
     DEBUGMSGTL(("helper:watcher:timestamp",
